Refactored main.js and decreased the frame rate.
Tidied up main.js and changed the delay argument from 25ms to 50ms in
setInterval(handleTick,25) within the function game.startTick(). The game
was running jerkily with only a 25ms delay. Presumably it was not enough
time to execute all functions in handleTick(). A 40ms delay gives a frame
rate of around 24 or 25 frames per second which is the standard for film.
So there is no need to go lower than 40ms really.